+++ src/qtt/measurements/post_processing/process_sawtooth_1d.py
@@ -1,8 +1,11 @@
-from qtt.measurements.post_processing.interfaces import SignalProcessorInterface
+from qilib.data_set import DataArray, DataSet
 
-from qilib.data_set import DataSet, DataArray
+from qtt.measurements.post_processing.interfaces import \
+    SignalProcessorInterface
+from qtt.utilities.tools import rdeprecated
 
 
+@rdeprecated(expire='1-1-2022')
 class ProcessSawtooth1D(SignalProcessorInterface):
 
     def run_process(self, signal_data: DataSet) -> DataSet: